training_pipeline/pyproject.toml
[tool.poetry]
name = "training-pipeline"
version = "0.1.0"
description = ""
authors = ["Angnami <mhtgoudjemi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
torch = "^2.2.1"
torchtext = "^0.17.1"
pandas = "^2.2.1"
numpy = "^1.26.4"
fire = "^0.6.0"
comet-ml = "^3.39.0"
python-dotenv = "^1.0.1"
unstructured = "^0.11.8"
beam-sdk = "^0.15.12"
transformers = "^4.38.2"
accelerate = "^0.28.0"
datasets = "^2.18.0"
kaggle = "^1.6.6"
pydantic = "^2.6.4"
scikit-learn = "^1.4.1.post1"
trl = "^0.8.6"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.2"
ipykernel = "^6.29.3"
notebook = "^7.1.1"
seaborn = "^0.13.2"
matplotlib = "^3.8.4"
wordcloud = "^1.9.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
